backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
neo4j>=5.0.0
python-dotenv>=1.0.0
